Internal REST Client updates and cleanup.
REST Calls will ALWAYS return it's results in the form of {Error:"reason", Data:{}) . On the go API we can now easily check for errors. (resp.Failed() == len(Error) > 0). UI can now also be integrated with confidence in API callbacks.

It doesn't matter if a STATUS is 200 or 500, the data returned will always have the same JSON format